requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
asttokens==3.0.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
datasets==3.1.0
decorator==5.1.1
dill==0.3.8
evaluate==0.4.3
executing==2.1.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
huggingface-hub==0.26.3
idna==3.10
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
jiwer==3.0.5
lightning-utilities==0.11.9
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
prompt_toolkit==3.0.48
propcache==0.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
PyAudio==0.2.14
Pygments==2.18.0
pyopenjtalk==0.3.4
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytorch-triton-rocm==3.1.0
pyttsx3==2.98
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
setuptools==75.6.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1+rocm6.2
torchaudio==2.5.1+rocm6.2
torchmetrics==1.6.0
torchvision==0.20.1+rocm6.2
tqdm==4.67.0
traitlets==5.14.3
transformers==4.47.0
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.18.3
